Add option to raise errors
By default, errors are just printed and ignored. The affected preview
thumbnails image is skipped. This is useful, when processing multiple
video files in a directory, because failure to create preview images for
individual files will not result in the remaining preview images not
being created.

Raising errors may be useful for debugging, however.